ad
Inform the user that if they do not include the second and third parameters to the constructor that it will add information of the executing script to the end of the URL to prevent this pass NULL and NULL for the last two parameters.

Also make it clear that $flags are the Constants and that you need to use a bitwise disjunction to include more than one. Also note that if the user is not adding the use statement to their file that they must pass the full name of the constant http\URL::STRIP_PASS or they will get exceptions

I had to look at this documentation and mess around for a few minutes to figure it out. I was only able to understand it because I've coded a lot of PHP I think the proposed changes make the documentation a little more clear and easier to understand. The window doesn't show any of the documentation about the methods or I would add to the __construct() part the information about flags are the constants and how to pass more than 1.